Avoid allocations in RecordingEventsAdapter
This tweaks the code to avoid allocating a closure on each poll, as well as a bytes buffer and sbe marshaller on each received message. The context is passed to the function callback by updating state inside RecordingEventsAdapter, and passing a (preallocated) method value to subscription poll.

In principle these could all be done at construction time, but since the RecordingEventsAdapter is public and didn't have a construction function before, it would not be backwards compatible to do so.

Ideally, we would also remove the (no longer needed) write to the global variable rangeChecking, but that is also technically not backwards compatible.