Reimplement FastMod3 so that it is actually fast
The previous implementation was primarily slow because the constant lookup table was encoded inside the function, which created a lot of extra work on each call.

That being said, the fast implementation is about 9 trivial operations and a table lookup, whereas the typical rendering of mod by constant is two multiplications (where one is trivial) and two trivial operations.

In fact, a modern go compiler will generate this transformation automatically, but we keep the expanded version to make sure this works with any older compiler as well (assuming the previous version written seven years ago was implemented in the complicated manner because the faster code wasn't automatically generated from the trivial implementation)