Fix Memcpy implementation performance
The implementation copied in chunks of eight bytes first, but instead of doing so until the final chunk, this just only did one eigth of the range in chunks of eight bytes.

Secondly, it did one fourth of the range in chunks of four bytes.

The updated code does all but the last 0-7 bytes in chunks of eight. If there is more than four bytes remaining it does one move in chunks of four, and the remaining 0-3 moves one byte at a time.