Expose DataFrameHeader constants as such
When wrapped into the structure, the golang compiler is unable to see them as constants and thus need to load the values from memory, and also cannot precompute anything with the value.

This introduces constants of the same name as the struct members, replacing the member access operator with an underscore. To make things backwards compatible, the old variables remain.