Replace uses of MakeBuffer with type-specific versions
The MakeBuffer helper method does reflection and type conversion under the hood, which introduces significant overhead for constructing a byte buffer meant for low-overhead processing.

This instead introduces NewBufferPointer and NewBufferSlice to use depending on the situation. For the slice case, this doesn't introduce any arguments for offset/length as that can be handled by slicing the slice before passing it in.